+++ com.alexander-yu-shamin.search-helper/Editor/Tools/UsedByTool.cs
@@ -37,7 +37,7 @@ namespace SearchHelper.Editor.Tools
             EGuiKit.Vertical(() => DrawVirtualScroll(windowRect, Contexts));
         }
 
-        public override void Run(Object selectedObject)
+        public override void Run(Object selectedObject, Settings settings)
         {
             if (selectedObject == null)
             {